# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ErrorResponse:
    properties:
      error: {}
      message:
        type: string
    type: object
...
  handlers.QRCodeData:
    properties:
      expires_at:
        type: string
      id_jadwal:
        type: integer
      jenis_sholat:
        type: string
      qr_code:
        type: string
      token:
        type: string
    type: object
  handlers.QRCodeResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.QRCodeData'
      message:
        type: string
    type: object
...
  handlers.VerifyQRData:
    properties:
      jenis_sholat:
        type: string
      jurusan:
        type: string
      kelas:
        type: string
      nama_siswa:
        type: string
      nis:
        type: string
      status:
        type: string
      tanggal:
        type: string
      valid:
        type: boolean
    type: object
  handlers.VerifyQRRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  handlers.VerifyQRResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.VerifyQRData'
      message:
        type: string
    type: object
...
info:
  contact:
    name: API Support
  description: API untuk sistem pencatatan absensi sholat siswa
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: API Sistem Absensi Sholat
  version: 1.0.0
paths:
...
  /qrcode/generate:
    get:
      consumes:
      - application/json
      description: Menghasilkan QR code untuk sesi absensi sholat aktif dan mengembalikannya
        sebagai data base64 dalam response JSON. QR code berisi token yang ter-sign
        dan berlaku selama 5 menit. Digunakan untuk menampilkan QR code di layar atau
        aplikasi mobile untuk siswa scan. Hanya untuk admin, guru, atau wali_kelas
      parameters:
      - description: Set to 'true' to force-generate QR for testing (non-production
          only)
        in: query
        name: force
        type: string
      - description: Specify prayer type (e.g., Maghrib) when forcing
        in: query
        name: jenis_sholat
        type: string
      - description: Specify jadwal ID when forcing
        in: query
        name: id_jadwal
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: QR code berhasil dibuat. Berisi base64 PNG, token, waktu kadaluarsa,
            jenis sholat, dan ID jadwal
          schema:
            $ref: '#/definitions/handlers.QRCodeResponse'
        "401":
          description: Tidak terotentikasi atau token tidak valid
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Akses ditolak - hanya admin, guru, atau wali_kelas yang dapat
            mengakses
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Tidak ada jadwal sholat aktif saat ini
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Kesalahan server internal saat membuat QR code
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Generate QR Code dengan format JSON (Base64)
      tags:
      - qrcode
  /qrcode/image:
    get:
      consumes:
      - application/json
      description: Menghasilkan dan mendownload QR code sebagai file PNG untuk sesi
        absensi sholat aktif. Mengembalikan file PNG 256x256 pixel dengan error correction
        level Medium. QR code berisi token yang ter-sign dan berlaku selama 5 menit.
        Gunakan endpoint ini untuk mencetak atau menampilkan QR code di layar. Hanya
        untuk admin, guru, atau wali_kelas
      parameters:
      - description: Set to 'true' to force-generate QR for testing (non-production
          only)
        in: query
        name: force
        type: string
      - description: Specify prayer type (e.g., Maghrib) when forcing
        in: query
        name: jenis_sholat
        type: string
      - description: Specify jadwal ID when forcing
        in: query
        name: id_jadwal
        type: integer
      produces:
      - image/png
      responses:
        "200":
          description: 'QR code PNG image (256x256 pixels). Filename: qrcode_[JenisSholat]_[IDJadwal].png'
          schema:
            type: file
        "401":
          description: Tidak terotentikasi atau token tidak valid
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Akses ditolak - hanya admin, guru, atau wali_kelas yang dapat
            mengakses
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Tidak ada jadwal sholat aktif saat ini
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Kesalahan server internal saat membuat QR code image
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download QR Code sebagai file PNG
      tags:
      - qrcode
  /qrcode/verify:
    post:
      consumes:
      - application/json
      description: 'Scan QR code yang sudah di-generate untuk mencatat kehadiran siswa
        dalam sesi absensi sholat. Endpoint ini memvalidasi token, memeriksa waktu
        kadaluarsa, dan mencatat absensi siswa dalam database. Token harus dalam format:
        Base64Payload.HexSignature. Hanya untuk siswa'
      parameters:
      - description: Token QR code hasil scan (dari endpoint generate atau image)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.VerifyQRRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Absensi berhasil dicatat. Menampilkan data siswa, jenis sholat,
            status kehadiran (hadir)
          schema:
            $ref: '#/definitions/handlers.VerifyQRResponse'
        "400":
          description: Token format tidak valid, format payload tidak valid, signature
            tidak valid, token sudah kadaluarsa (lebih dari 5 menit), atau QR code
            tidak valid untuk hari ini
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Tidak terotentikasi atau NIS tidak ditemukan di token JWT
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Akses ditolak - hanya siswa yang dapat mengakses
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Siswa atau jadwal sholat tidak ditemukan
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Konflik - siswa sudah absen untuk jadwal sholat ini pada hari
            yang sama
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Kesalahan server internal saat mencatat absensi
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Scan dan verifikasi QR Code untuk mencatat absensi
      tags:
      - qrcode
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"